refactor(desktop): typography settings as two always-visible surface sections
The Typography section had one shared control card whose target surface
was switched by clicking the Editor/Terminal preview headers below it,
which was easy to miss and confusing. Split it into self-contained
Editor and Terminal sections — each with its own controls, reset, and
live preview — rendered as divider-separated rows in one bordered
container (ThemeSection pattern), with the fine-tune fields inline
instead of behind a collapsible.
