+++ apps/desktop/src/renderer/routes/_authenticated/settings/appearance/components/AppearanceSettings/components/FontSettingSection/components/TypographySurfaceCard/index.ts
@@ -1,0 +1,4 @@
+export {
+	type FontSettingsUpdate,
+	TypographySurfaceCard,
+} from "./TypographySurfaceCard";
